Remove circular bean dependency in tests.
